pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["NatanDuarte <natanprocopio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
httpx = "^0.24.1"
uvicorn = "^0.23.2"
pytest = "^7.4.2"
pymongo = "^4.5.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
httpie = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
